backend_fastapi/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
ccxt>=4.2.0
finance-datareader>=0.9.60
yfinance>=0.2.36
fredapi>=0.4.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
